Cope with existing S3 bucket in non-default region
This is a functional fix for the problem, but Fog still complains about
following redirects, since we don't support multiple regions across the
board.

This problem is only reproducible with real S3 backend, not the
Fog::Mock.

This is despite https://github.com/fog/fog/pull/2461 - maybe the
mock would need further modification to support the way we're
instantiating Fog::Storage.

[Fixes #85081296]