Implement a CachingGitAccessor
Removes the need for conditional code higher up the stack.

The presence of this class assumes that we never want to clone twice for
the same repo and target directory, in certain situations.

That might not always be a safe assumption.

Remove a test that was testing caching, as it was coupled to the
previous implementation. As long as we select the right type, we should
be safe without that test.

[#85024596]