Pass command instances to command runner
Previously, we were passing command classes.

This was encouraging us to inherit from Bookbinder::Command, which is really just providing a
constructor that doesn't apply to all commands.

Passing classes meant that the command runner was hard to test. If we
pass instances, we can easily test the command runner with test doubles.

[#86808160]