Pass target_ref to copy method not ctor
GitHubRepository was asking for a target_ref for clones in its
constructor. In many cases this is unnecessary.

This change forces clients of the GHRepo to do their own processing of
the refs.

Now that the processing of refs is in Bind's hands, it can decide
whether to pass 'master' based on whether the user has asked for
configured refs to be ignored.

[#86816248]
[#89023704]