Revert to using only toc_path in config.yml
Previously, we had users specify both base_path and toc_path in the
config.yml to help determine the relative link resolution while parsing
the TOC pages in nav generation. This can instead be handled by working
only with the path to the TOC file, so this commit simplifies the
config.yml accordingly.

[#100752434]