update spec to match template_app values
Previous commits to `search.yml` in template_app added h2 and h3 as fields with
weights, changed weight of title from ^10 to ^20, and changed the
minimum_should_match parameter from 1 (return pages that match at least
1 search term) with 75% (return pages where at least 75% of the entered
terms match). This commit updates spec to match the changes in the
previous commits.