support fuzzy matching of directories when using non `master` directory structure
The current workflow for `yield_for_code_snippet` assumes all snippets are always in the `master` branch.
The lookup path for the snippet directory would be (in order):
- repo-name
- repo-name-branch-name
- repo-name-org-name-branch-name

When we moved from `master` to `develop`/`release` pattern, our document generation starting failing in the Concourse pipeline.
This is because our input name was `platform-automation-develop` instead of the assumed `platform-automation-master` as noted above.

This change supports the fact that inputs (from the generated pipeline), do not support that format.
It will perform precedence lookup, as above, and then do a _best match_ for a directory.
This means that our workflow of `develop`/`release` will be supported and we do not break existing `master` patterns.
[#161320648]

Signed-off-by: JT Archie <jarchie@pivotal.io>