Update template_app dependencies to be compatible w/ cflinuxfs3
- Use Ruby 2.3.8 (cflinuxfs3 stack buildpacks only include 2.3.7,
2.3.8). We will soon need to change this again as 2.3.x is being EOL'd
in a few months.
- Use 3.12.0 puma (3.6.0 did not work with the new buildpack).
- Use 2.0.2 elasticsearch (2.0.0 is required for the new buildpack, go
to latest).
- It is unfortunate that the template_app has a lockfile -- this limits
our flexibility with bookbinder app outputs. Removing the lockfile
*could* gain us some flexibility, but we might have to think carefully
about the version locks in the Gemfile. I suspect the lockfile was
originally checked in because bookbinder had continuous support and
version bumps were cheap. This is no longer the case, so a strategy
change may be in order.