Update to Ruby 2.5.3
- Move to later versions so bookbinder can be used for longer without
changes (2.3.x line is being EOLed in March 2019).
- Remove stabby lambda syntax. This is using a undoc'd feature of detect that
allows one to return a default if nothing is detected. I suspect you
could also || the detect with the desired default and this might be more
readable but there weren't tests (afaict) around the defaulting behavior
so I chose to make as few changes as possible.
- Apparently FileUtils.rmdir used to rescue the "directory not empty"
error, so we never actually deleted the directory. Switch to rm_rf which
should delete the dir. It sounds like there is a security issue when
using with directories that may have unknown contents (like symlinks,
etc) but that doesn't seem to be a concern here (plus is in test files).